Refactor to functional component
